+++ packages/apollo-server-core/src/ApolloServer.ts
@@ -627,8 +627,8 @@ export class ApolloServerBase {
     this.subscriptionServer = SubscriptionServer.create(
       {
         schema,
-        execute,
-        subscribe,
+        execute: this.config.executeFn || execute,
+        subscribe: this.config.subscribeFn || subscribe,
         onConnect: onConnect
           ? onConnect
           : (connectionParams: Object) => ({ ...connectionParams }),
@@ -837,6 +837,8 @@ export class ApolloServerBase {
       documentStore,
       extensions,
       context,
+      executeFn: this.config.executeFn,
+      subscribeFn: this.config.subscribeFn,
       // Allow overrides from options. Be explicit about a couple of them to
       // avoid a bad side effect of the otherwise useful noUnusedLocals option
       // (https://github.com/Microsoft/TypeScript/issues/21673).