+++ packages/apollo-server-core/src/runHttpQuery.ts
@@ -167,6 +167,7 @@ export async function runHttpQuery(
     validationRules: options.validationRules,
     executor: options.executor,
     fieldResolver: options.fieldResolver,
+    executeFn: options.executeFn,
 
     // FIXME: Use proper option types to ensure this
     // The cache is guaranteed to be initialized in ApolloServer, and