+++ packages/apollo-server-core/CHANGELOG.md
@@ -13,3 +13,4 @@
 * `apollo-server-core`: context creation can be async and errors are formatted to include error code [PR #1024](https://github.com/apollographql/apollo-server/pull/1024)
 * `apollo-server-core`: add `mocks` parameter to the base constructor(applies to all variants) [PR#1017](https://github.com/apollographql/apollo-server/pull/1017)
 * `apollo-server-core`: Remove printing of stack traces with `debug` option and include response in logging function[PR#1018](https://github.com/apollographql/apollo-server/pull/1018)
+* `apollo-server-core`: Custom `execute` and `subscribe` functions [PR #4167](https://github.com/apollographql/apollo-server/pull/4167)